move _have_dnspython
